refactor(subagents): extract model resolution into subagent-model.ts
Move the subagent model-resolution logic out of the oversized manager.ts
into a self-contained src/agent/subagents/subagent-model.ts:
getModelHandleFromAgent, getPrimaryAgentModelHandle, getCurrentBillingTier,
resolveSubagentModel, and the provider/inherit helpers.

Pure move, no behavior change. The module depends only on lower-level
model/backend helpers and is not imported back by manager (no cycle).
manager.ts and the resolution test now import from the new module directly
(no re-export barrel), and the manager.ts size baseline is ratcheted down
1577 -> 1357.
