+++ src/agent/subagent-model-resolution.test.ts
@@ -8,12 +8,14 @@ import {
 import {
   buildSubagentArgs,
   buildSubagentPrompt,
-  getModelHandleFromAgent,
   recallPromptForBackend,
   resolveSubagentLauncher,
-  resolveSubagentModel,
   resolveSubagentWorkingDirectory,
 } from "@/agent/subagents/manager";
+import {
+  getModelHandleFromAgent,
+  resolveSubagentModel,
+} from "@/agent/subagents/subagent-model";
 
 describe("recallPromptForBackend", () => {
   test("uses separate API and local recall prompts", () => {
